Rename module to disable_doctest
